Add files via upload
Fun little asteroids game I created in vanilla JavaScript.  No libraries/frameworks.  I create a simple engine to handle the game loop, drawing/rendering, objects, collision and physics.  This engine is the basis (or essentially the skeleton) for the one I plan on creating with MonoGame/Nez or libGDX upon which I plan to build and release games under OZsoft.  This is the starting point.